+++ src/ej_conversations/urls/public_conversations.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from ..views import PublicConversationView
+from .. import views
 
 app_name = "ej_conversations"
 conversation_url = "<int:conversation_id>/<slug:slug>/"
@@ -7,7 +7,22 @@ conversation_url = "<int:conversation_id>/<slug:slug>/"
 urlpatterns = [
     path(
         "",
-        PublicConversationView.as_view(),
+        views.PublicConversationView.as_view(),
         name="list",
+    ),
+    path(
+        "tags/",
+        views.ConversationsFilterByTag.as_view(),
+        name="filter-tags",
+    ),
+    path(
+        "tags/user/",
+        views.UserConversationsFilterByTag.as_view(),
+        name="filter-tags-user",
+    ),
+    path(
+        "tags/promoted/",
+        views.PromotedConversationsFilterByTag.as_view(),
+        name="filter-promoted",
     ),
 ]