+++ src/ej_users/api.py
@@ -68,6 +68,9 @@ class TokenViewSet(viewsets.ViewSet):
 
     @action(detail=False, methods=["post"], permission_classes=[AllowAny])
     def token(self, request):
+        """
+        Returns an access_token and refresh_token for an user.
+        """
         serializer = UserAuthSerializer(data=request.data)
         if not serializer.is_valid():
             return Response(serializer.errors, status=400)